Work around reverse for in logic in coffeescript
The counting backwards in a range wassn't working correctly, it wasn't
putting the string together in reverse order resulting in commas not
being in the right place. Switching to the while i-- ensure that the
right order is being used.

Also moved the call to check if a comma is needed to after the number
has been added. This makes the counts of 3 work.